+++ lib/countly.js
@@ -218,14 +218,7 @@ Countly.Bulk = Bulk;
      * WARNING!!!
      * Use only for testing purposes!!!
      * 
-     * Resets Countly to its initial state (used mainly to wipe the queues in memory) and must be coupled with Countly.init to change configuration:
-     * //initial initialization
-     * Countly.init({....});
-     * ...
-     * //wipe the state but not config
-     * Countly.halt();
-     * //New init with new config
-     * Countly.init({....});
+     * Resets Countly to its initial state (used mainly to wipe the queues in memory)
      */
     Countly.halt = function name() {
         cc.log(cc.logLevelEnums.WARNING, "halt, Resetting Countly");
