+++ test/test-internal-limits.js
@@ -1,10 +1,24 @@
 /* eslint-disable no-console */
 /* global describe, it, */
 var assert = require("assert"),
+    Countly = require("../lib/countly.js"),
     hp = require("./helpers/helper-functions"),
     cc = require("../lib/countly-common");
 
-//run with: mocha --timeout 20000 .\test\test-internal-limits.js --exit , at your CLI
+//standard init for tests
+function initLimitsMain() {
+    Countly.init({
+        app_key: "YOUR_APP_KEY",
+        url: "https://try.count.ly",
+        max_events: -1,
+        max_key_length: 8, //set maximum key length here
+        max_value_size: 8, //set maximum value length here
+        max_segmentation_values: 3, //set maximum segmentation number here
+        max_breadcrumb_count: 2, //set maximum number of logs that will be stored before erasing old ones
+        max_stack_trace_lines_per_thread: 3, //set maximum number of lines for stack trace
+        max_stack_trace_line_length: 10 //set maximum length of a line for stack trace
+    });
+}
 
 
 
@@ -47,9 +61,19 @@ describe("Testing internal limits", function() {
     //clear storage
         hp.clearStorage();
         //init Countly
-        hp.initLimitsMain();
+        initLimitsMain();
         //send event
-        hp.customLimitsEvent();
+        Countly.add_event({
+            key: "Enter your key here",
+            count: 1,
+            segmentation: {
+                "key of 1st seg": "Value of 1st seg",
+                "key of 2nd seg": "Value of 2nd seg",
+                "key of 3rd seg": "Value of 3rd seg",
+                "key of 4th seg": "Value of 4th seg",
+                "key of 5th seg": "Value of 5th seg",
+            },
+        });
         setTimeout(() => {
             //read event queue
             var event = hp.readEventQueue()[0];
@@ -68,9 +92,9 @@ describe("Testing internal limits", function() {
         //clear storage
         hp.clearStorage();
         //init Countly
-        hp.initLimitsMain();
+        initLimitsMain();
         //page view
-        hp.trackLimitsPageView();
+        Countly.track_pageview("a very long page name");
         //test
         setTimeout(() => {
             //read event queue
@@ -89,11 +113,20 @@ describe("Testing internal limits", function() {
         //clear storage
         hp.clearStorage();
         //init Countly
-        hp.initLimitsMain();
+        initLimitsMain();
         //add log
-        hp.addLog();
+        Countly.add_log("log1");
+        Countly.add_log("log2");
+        Countly.add_log("log3");
+        Countly.add_log("log4");
+        Countly.add_log("log5 too many");
+        Countly.add_log("log6");
+        Countly.add_log("log7");
         //and log error to see them all
-        hp.errorLog();
+        var error = {
+            stack: "Lorem ipsum dolor sit amet,\n consectetur adipiscing elit, sed do eiusmod tempor\n incididunt ut labore et dolore magna\n aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.\n Duis aute irure dolor in reprehenderit in voluptate\n velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia\n deserunt mollit anim id\n est laborum.",
+        };
+        Countly.log_error(error);
         //test
         setTimeout(() => {
             //read event queue
@@ -123,9 +156,27 @@ describe("Testing internal limits", function() {
         //clear storage
         hp.clearStorage();
         //init Countly
-        hp.initLimitsMain();
+        initLimitsMain();
         //add user details
-        hp.userLimitsDetails();
+        Countly.user_details({
+            name: "Gottlob Frege",
+            username: "Grundgesetze",
+            email: "test@isatest.com",
+            organization: "Bialloblotzsky",
+            phone: "+4555999423",
+            //Web URL pointing to user picture
+            picture:
+            "https://ih0.redbubble.net/image.276305970.7419/flat,550x550,075,f.u3.jpg",
+            gender: "M",
+            byear: 1848, //birth year
+            custom: {
+                "SEGkey 1st one": "SEGVal 1st one",
+                "SEGkey 2st one": "SEGVal 2st one",
+                "SEGkey 3st one": "SEGVal 3st one",
+                "SEGkey 4st one": "SEGVal 4st one",
+                "SEGkey 5st one": "SEGVal 5st one",
+            },
+        });
         //test
         setTimeout(() => {
             //read event queue
@@ -159,9 +210,18 @@ describe("Testing internal limits", function() {
         //clear storage
         hp.clearStorage();
         //init Countly
-        hp.initLimitsMain();
+        initLimitsMain();
         //add custom properties
-        hp.userLimitsData();
+        Countly.userData.set("name of a character", "Bertrand Arthur William Russell"); //set custom property
+        Countly.userData.set_once("A galaxy far far away", "Called B48FF"); //set custom property only if property does not exist
+        Countly.userData.increment_by("byear", 123456789012345); //increment value in key by provided value
+        Countly.userData.multiply("byear", 2345678901234567); //multiply value in key by provided value
+        Countly.userData.max("byear", 3456789012345678); //save max value between current and provided
+        Countly.userData.min("byear", 4567890123456789); //save min value between current and provided
+        Countly.userData.push("gender", "II Fernando Valdez"); //add value to key as array element
+        Countly.userData.push_unique("gender", "III Fernando Valdez"); //add value to key as array element, but only store unique values in array
+        Countly.userData.pull("gender", "III Fernando Valdez"); //remove value from array under property with key as name
+        Countly.userData.save();
         //test
         setTimeout(() => {
             //read event queue
